f123

+++ modules/queue/include/queue.h
@@ -5,14 +5,13 @@
 
 #include <iostream>
 #include <vector>
-#include <cstdio>
 
 class Queue {
  private:
     int tail_;
     int head_;
     std::vector<int> pq_;
-    size_t size_;
+    int size_;
     int count_;
 
     int GetNextIndex(int index) const;
